+++ src/main.rs
@@ -12,7 +12,7 @@ fn main() {
     let camera = Camera::with_transform(
         600,
         600,
-        std::f32::consts::FRAC_PI_3,
+        std::f64::consts::FRAC_PI_3,
         view_transform(
             &Tuple::new_point(0., 1.5, -5.),
             &Tuple::new_point(0., 1., 0.),
@@ -48,15 +48,15 @@ fn build_world() -> World {
     );
     let left_wall = Sphere::new(
         translate(0., 0., 5.)
-            * rotate_y(-std::f32::consts::FRAC_PI_4)
-            * rotate_x(std::f32::consts::FRAC_PI_2)
+            * rotate_y(-std::f64::consts::FRAC_PI_4)
+            * rotate_x(std::f64::consts::FRAC_PI_2)
             * scale(10., 0.01, 10.),
         floor.material.clone(),
     );
     let right_wall = Sphere::new(
         translate(0., 0., 5.)
-            * rotate_y(std::f32::consts::FRAC_PI_4)
-            * rotate_x(std::f32::consts::FRAC_PI_2)
+            * rotate_y(std::f64::consts::FRAC_PI_4)
+            * rotate_x(std::f64::consts::FRAC_PI_2)
             * scale(10., 0.01, 10.),
         floor.material.clone(),
     );
