+++ src/tuple.rs
@@ -6,22 +6,22 @@ use std::ops::{Add, Div, Mul, Neg, Sub};
 // TODO: SIMD
 #[derive(Debug, PartialEq, Clone)]
 pub struct Tuple {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
-    pub w: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
+    pub w: f64,
 }
 
 impl Tuple {
-    pub fn new(x: f32, y: f32, z: f32, w: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64, w: f64) -> Self {
         Self { x, y, z, w }
     }
 
-    pub fn new_point(x: f32, y: f32, z: f32) -> Self {
+    pub fn new_point(x: f64, y: f64, z: f64) -> Self {
         Self::new(x, y, z, 1.)
     }
 
-    pub fn new_vector(x: f32, y: f32, z: f32) -> Self {
+    pub fn new_vector(x: f64, y: f64, z: f64) -> Self {
         Self::new(x, y, z, 0.)
     }
 
@@ -60,7 +60,7 @@ impl Tuple {
             && is_close(self.z, rhs.z)
     }
 
-    pub fn magnitude(&self) -> f32 {
+    pub fn magnitude(&self) -> f64 {
         (self.dot(self)).sqrt()
     }
 
@@ -75,7 +75,7 @@ impl Tuple {
         }
     }
 
-    pub fn dot(&self, rhs: &Self) -> f32 {
+    pub fn dot(&self, rhs: &Self) -> f64 {
         // maybe assert_eq!(self.w, rhs.w, 0.)
         self.x * rhs.x + self.y * rhs.y + self.z * rhs.z
     }
@@ -136,10 +136,10 @@ impl Neg for Tuple {
     }
 }
 
-impl Mul<f32> for Tuple {
+impl Mul<f64> for Tuple {
     type Output = Self;
 
-    fn mul(self, scalar: f32) -> Self {
+    fn mul(self, scalar: f64) -> Self {
         Self {
             x: self.x * scalar,
             y: self.y * scalar,
@@ -149,7 +149,7 @@ impl Mul<f32> for Tuple {
     }
 }
 
-impl Mul<Tuple> for f32 {
+impl Mul<Tuple> for f64 {
     type Output = Tuple;
 
     fn mul(self, tuple: Tuple) -> Tuple {
@@ -162,10 +162,10 @@ impl Mul<Tuple> for f32 {
     }
 }
 
-impl Div<f32> for Tuple {
+impl Div<f64> for Tuple {
     type Output = Self;
 
-    fn div(self, divisor: f32) -> Self {
+    fn div(self, divisor: f64) -> Self {
         self * (1. / divisor)
     }
 }
@@ -266,11 +266,11 @@ mod tests {
         assert_eq!(Tuple::new_vector(0., 3., 4.).magnitude(), 5.);
         assert_eq!(
             Tuple::new_vector(1., 2., 3.).magnitude(),
-            (14. as f32).sqrt()
+            (14. as f64).sqrt()
         );
         assert_eq!(
             Tuple::new_vector(1., -2., -3.).magnitude(),
-            (14. as f32).sqrt()
+            (14. as f64).sqrt()
         );
     }
 
@@ -283,7 +283,7 @@ mod tests {
         assert!(Tuple::new_vector(3., 4., 0.)
             .normalize()
             .is_close(&Tuple::new_vector(0.6, 0.8, 0.)));
-        let sqrt14 = (14. as f32).sqrt();
+        let sqrt14 = (14. as f64).sqrt();
         assert!(Tuple::new_vector(1., -2., -3.)
             .normalize()
             .is_close(&Tuple::new_vector(1. / sqrt14, -2. / sqrt14, -3. / sqrt14)));
@@ -320,7 +320,7 @@ mod tests {
         assert!(Tuple::new_vector(1., -1., 0.)
             .reflect(&Tuple::up())
             .is_close(&Tuple::new_vector(1., 1., 0.)));
-        let sqrt_half = std::f32::consts::FRAC_1_SQRT_2;
+        let sqrt_half = std::f64::consts::FRAC_1_SQRT_2;
         assert!(Tuple::new_vector(0., -1., 0.)
             .reflect(&Tuple::new_vector(sqrt_half, sqrt_half, 0.))
             .is_close(&Tuple::new_vector(1., 0., 0.)));
